requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.4.0
numpy==2.3.2
pandas==2.3.1
PyMySQL==1.1.1
pyodbc==5.2.0
pyRFC3339==2.0.1
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.42
typing_extensions==4.14.1
tzdata==2025.2
Werkzeug==3.1.3
